pyproject.toml
[project]
name = "drf-sandbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-rest-framework>=0.1.0",
    "django-telegram-login>=0.2.3",
    "djangorestframework-simplejwt>=5.5.1",
    "icecream>=2.1.10",
    "requests>=2.32.5",
]
